feat: Moved node resources ui into it's own component
