+++ packages/playground/src/components/node_resources_charts.vue
@@ -15,21 +15,7 @@
       </v-col>
     </v-row>
 
-    <v-row justify="center">
-      <div v-for="item in resources" :key="item.name" class="mx-6 d-flex flex-column pt-2 mt-2 align-center">
-        <div class="mb-2">{{ item.name }}</div>
-        <div class="text-center">
-          <v-progress-circular
-            :model-value="item.value !== 'NaN' ? item.value : 0"
-            :size="150"
-            :width="15"
-            color="info"
-            :indeterminate="indeterminate"
-            >{{ item.value !== "NaN" ? item.value + "%" : "N/A" }}
-          </v-progress-circular>
-        </div>
-      </div>
-    </v-row>
+    <NodeResources :node="node" />
 
     <v-row justify="center">
       <div class="d-flex my-6 align-center justify-center">
@@ -53,7 +39,10 @@ import type { ResourceWrapper } from "@/types";
 import { GrafanaStatistics } from "@/utils/get_metrics_url";
 import { getNodeStatusColor } from "@/utils/get_nodes";
 
+import NodeResources from "./node_resources.vue";
+
 export default {
+  components: { NodeResources },
   props: {
     node: {
       type: Object as PropType<GridNode>,
@@ -68,12 +57,8 @@ export default {
       required: true,
     },
   },
-  async mounted() {
-    this.resources = await this.getNodeResources();
-  },
   setup(props) {
     const resources = ref<ResourceWrapper[]>([]);
-    const renamedResources = ["CPU", "SSD", "HDD", "RAM"];
     const loading = ref<boolean>(false);
     const indeterminate = ref<boolean>(false);
     const getNodeHealthUrl = async () => {
@@ -82,33 +67,11 @@ export default {
       window.open(nodeHealthUrl, "_blank");
     };
 
-    const getNodeResources = async () => {
-      loading.value = true;
-      return ["cru", "sru", "hru", "mru"].map((i, idx) => {
-        let value;
-        if (props.node.stats && props.node.stats.system) {
-          value =
-            ((Reflect.get(props.node.stats.used, i) + Reflect.get(props.node.stats.system, i)) /
-              Reflect.get(props.node.stats.total, i)) *
-            100;
-        } else {
-          value = (Reflect.get(props.node.used_resources, i) / Reflect.get(props.node.total_resources, i)) * 100;
-        }
-        loading.value = false;
-        return {
-          id: idx + 1,
-          value: value.toFixed(2),
-          name: renamedResources[idx],
-        };
-      });
-    };
-
     return {
       NodeStatus,
       resources,
       loading,
       indeterminate,
-      getNodeResources,
       getNodeHealthUrl,
       getNodeStatusColor,
     };
